% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{csm_arr_fr_active}
\alias{csm_arr_fr_active}
\title{Fraction of active enzymes based on modified Arrhenius function}
\usage{
csm_arr_fr_active(Tt, H, E, To)
}
\arguments{
\item{Tt}{temperature in Celsius}

\item{H}{deactivation energy parameter}

\item{E}{activation energy parameter}

\item{To}{optimum temperature in Celsius}
}
\value{
a numeric value of the fraction of active enzymes rate at temperature Tt
}
\description{
This function computes the fraction of active enzymes according to the
the modified Arrhenius function. The fraction of denatured enzymes can be
calculated by subtracting this function from 1.
}
