% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{csm_boltz_max}
\alias{csm_boltz_max}
\title{Calculate the Boltzmann operator smooth approximation to the maximum function}
\usage{
csm_boltz_max(x1, x2, alpha = 1)
}
\arguments{
\item{x1}{a scalar numeric input value}

\item{x2}{a scalar numeric input value}

\item{alpha}{a scalar numeric value providing the sharpness of the approximation function}
}
\value{
a numeric value
}
\description{
Calculate the Boltzmann operator smooth approximation to the maximum function
}
