% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{csm_calc_logistic_r}
\alias{csm_calc_logistic_r}
\title{Calculate logistic rate parameter}
\usage{
csm_calc_logistic_r(x, f_x, x0)
}
\arguments{
\item{x}{a numeric value providing a target input value}

\item{f_x}{a numeric value providing the target logistic function value at x}

\item{x0}{a numeric value providing the mid-point parameter of the logistic function}
}
\value{
a numeric value
}
\description{
Calculate logistic rate parameter
}
