% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{csm_calc_n_up}
\alias{csm_calc_n_up}
\title{Hill coefficient for up-regulation}
\usage{
csm_calc_n_up(x, L, K)
}
\arguments{
\item{x}{a numeric value providing the value of the Hill equation at L}

\item{L}{a numeric value providing the ligand concentration}

\item{K}{a numeric value providing the ligand concentration at half occupation}
}
\value{
a numeric value
}
\description{
Hill coefficient for up-regulation
}
