% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{csm_log_sum_exp}
\alias{csm_log_sum_exp}
\title{Calculate the LogSumExp smooth approximation to a bounded variable}
\usage{
csm_log_sum_exp(x, y, alpha = 1)

csm_log_sum_exp(x, y, alpha = 1)
}
\arguments{
\item{alpha}{a numeric value providing the sharpness of the approximation function}

\item{x1}{a numeric input value}

\item{x2}{a numeric input value}
}
\value{
a numeric value

a numeric value
}
\description{
Calculate the LogSumExp smooth approximation to a bounded variable

Calculate the LogSumExp smooth approximation to a bounded variable
}
