% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{csm_logistic}
\alias{csm_logistic}
\title{Logistic switch function}
\usage{
csm_logistic(x, r, x0)
}
\arguments{
\item{x}{a numeric value providing the input variable}

\item{r}{a numeric value providing the rate parameter of the logistic function}

\item{x0}{a numeric value providing the mid-point parameter of the logistic function}
}
\value{
a numeric value between 0 and 1
}
\description{
Logistic switch function
}
