% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_run_sim.R
\name{csm_run_sim}
\alias{csm_run_sim}
\title{Run a Cropping System Model (CSM) simulation}
\usage{
csm_run_sim(model_function, y_init, t, events, ..., method = "euler")
}
\arguments{
\item{model_function}{a rendered model produced by
\code{\link[=csm_render_model]{csm_render_model()}}}

\item{y_init}{a vector of initial values for the model state variables}

\item{t}{an optional vector of time points for which simulated model outputs
are desired}

\item{events}{a data frame of external events with a column named \code{time}
that contains the time of each event and a list column named \code{event}
each element of which is a function that accepts a vector of state variables
and returns a vector of state variables to be used as the initial conditions
for continuing the simulation after the event. The times for all events must
be within the time range specified by the \code{t} argument.}

\item{...}{additional arguments to pass to model_function for simulation}

\item{method}{numerical integration method to be used. See \code{\link[deSolve:ode]{deSolve::ode()}}
for more details}
}
\value{
a data frame with one row for each time point specified by \code{t}
}
\description{
Run a Cropping System Model (CSM) simulation
}
\examples{

# Define state variables
lv_state <- csm_create_state(
  c("x", "y"),
  definition = c("prey", "predator"),
  units = c("rabbits per square km", "foxes per square km"),
  expression(~alpha*x-beta*x*y, ~delta*x*y-gamma*y))

# Define parameters
lv_parameters <- csm_create_parameter(
  c("alpha", "beta", "gamma", "delta"),
  definition = c("maximum prey per capita growth rate",
                 "effect of predator population on prey death rate",
                 "predator per capita death rate",
                 "effect of prey population on predator growth rate"),
  units = c("rabbits per rabbit", "per fox",
            "foxes per fox", "foxes per rabbit"))

# Define model
lotka_volterra_model <-
  csm_create_model(
    state = lv_state,
    parms = lv_parameters)

# Render model into a callable R function
lotka_volterra_fun <-
  csm_render_model(lotka_volterra_model,
                   output_type = "function",
                   language = "R")

# Run model simulation
lotka_volterra_out <-
  csm_run_sim(model_function = lotka_volterra_fun,
              y_init = c(x = 10,
                         y = 10),
              t = csm_time_vector(0, 100, 0.01),
              parms = c(alpha = 1.1,
                        beta = 0.4,
                        gamma = 0.1,
                        delta = 0.4))

}
