% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{csm_smooth_bounds}
\alias{csm_smooth_bounds}
\title{Calculate smooth approximation to a bounded variable}
\usage{
csm_smooth_bounds(x, min_val, max_val)
}
\arguments{
\item{x}{a numeric input value}

\item{min_bound}{a numeric value of the minimum bound}

\item{max_bound}{a numeric value of the maximum bound}
}
\value{
a numeric value
}
\description{
Calculate smooth approximation to a bounded variable
}
